+++ CHANGES.md
@@ -25,6 +25,7 @@ Changes in 11.4
 - Automatically clear Access Log entries older than one year in AccessLog.php
 - Only list events for selected month by default in Calendar.php
 - Use DBUpsert() in Comments.inc.php
+- Set Moodle course end date (enddate parameter) in plugins/Moodle/Scheduling/Course.php
 
 Changes in 11.3.3
 -----------------